Facets: Fix logic for disabled facets
This removes the logic that treats single choice facets where results===futureResults special. The assumption was, that this case is an indicator for a facet that should be inactive. However, we only looked at total results, not at the actual content of the results (which we cannot do).

It also improves the tooltips for all facets.

And the styling for the icon buttons.