ReportPage: Introduce TOC
Add sticky TOC on the left; add TOC-hashes to Headlines.

I did not make this super dynamic (for now), just to keep it simple.

Note: The hash URL parts are not translated in the original version of the page, so I left them this way to not break any existing URLs.