Fix Link comp wrapped inside headless UI with unused forwardRef
Using our custom Link component inside a headless UI Menu.Item produces an error:

>

The error does not hint at headless UI as the source of the issue, but it looks like it is. Wrapping our Link component with React.forwardRef does solved the problem (as described in https://github.com/tailwindlabs/headlessui/issues/1512#issuecomment-1139937513). However, we do not use the given ref; or at least the Gatsby Link component does not "want" it as ref={ref}, nor as innerRef.

Debugging this was way too painful … :-/